Fix non incrementing bug
simpleSync is a struct referenced to increment
a specific instance of integer a number of times
at a Later point in execution.

It was not incrementing because the method call
generated the nil version of the struct each time.
The fix was simply to change the struct references
to pointers all around.

Now the correct struct is used during runtime