Add Test for unexported clusterTogether()
Try new testing strategy:
Where step-wise and unit tests for pkg function are wanted.
Include in the same pkg.

Downside? Tests are included for the exported API.
Upside? Internals are tested in a whitebox style.
Using a pkg_test instead would leverage the blackbox style.
The real question? Is this a unit test? Debate GO!